requirements.txt
streamlit==1.39.0
pandas==2.2.3
gspread==6.1.4
oauth2client==4.1.3
streamlit-lottie==0.0.5
python-dotenv
google-auth
google-auth-oauthlib

